+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- web: `list_languages` now uses reForis translations instead of Foris translations to determine available localization,  
+    since Foris is no longer present in TOS 5.3.0
+    
 ### Fixed:
 - networks: fix reading multiple interfaces of wan (e.g. bridge)